feat: :+1: organize sample_item files
