+++ lib/src/app.dart
@@ -7,10 +7,10 @@ import 'package:flutter_localizations/flutter_localizations.dart';
 
 // Project imports:
 import 'package:yomutube/src/feature/authentication/presentation/login_view.dart';
+import 'package:yomutube/src/feature/sample/presentation/sample_item_details_view.dart';
+import 'package:yomutube/src/feature/sample/presentation/sample_item_list_view.dart';
 import 'package:yomutube/src/feature/settings/presentation/settings_controller.dart';
 import 'package:yomutube/src/feature/settings/presentation/settings_view.dart';
-import 'package:yomutube/src/sample_feature/sample_item_details_view.dart';
-import 'package:yomutube/src/sample_feature/sample_item_list_view.dart';
 
 /// The Widget that configures your application.
 class MyApp extends StatelessWidget {
