+++ lib/src/feature/sample/presentation/component/sample_item.dart
