+++ lib/src/feature/sample/presentation/sample_item_details_view.dart
